--- exercices/6e/6C10.js
@@ -1,13 +1,0 @@
-import Tables from './6C101.js';
-
-/**
- * Tables de multiplications classiques, à trou ou un mélange des deux.
- *
- * Par défaut ce sont les tables de 2 à 9 mais on peut choisir les tables que l'on veut
- * @Auteur Rémi Angot
- * Référence 6C10-1
- */
-export default function Tables_parametres(tables_par_defaut = "2-3-4-5-6-7-8-9"){
-    Tables.call(this,tables_par_defaut)
-    this.sup2 = 2;
-}
