+++ exercices/6e/6G20-2.js
@@ -5,8 +5,8 @@ import Vocabulaire_des_triangles from '../5e/5G20-1.js'
  * @author Sébastien Lozano
  */
 export default function Vocabulaire_des_triangles_6e() {
-  this.classe = 6;
   Vocabulaire_des_triangles.call(this);
   this.titre = "Vocabulaire des triangles";
+  this.classe = 6;
 }
