+++ exercices/6e/6G21-1.js
@@ -1,5 +1,6 @@
 import Construire_un_triangle_avec_cible from './_Construire_un_triangle_avec_cible.js';
 export default function Construire_un_triangle_avec_cible_6e() {
   Construire_un_triangle_avec_cible.call(this);
+  this.titre="Construire un triangle avec cible";
   this.classe = 6;
 }
