+++ exercices/6e/_6CM01.js
