+++ modules/menuDesExercicesDisponibles.js
@@ -1,6 +1,7 @@
 import { tridictionnaire, filtreDictionnaire }  from "./outils.js" ;
-export let dictionnaireDesExercices = {"4G53":{"url":"/exercices/4e/4G53.js","titre":"Calculs de volumes"},"4C30":{"url":"/exercices/4e/4C30.js","titre":"Puissances de 10 : Le sens des règles de calculs"},"4C10-2":{"url":"/exercices/4e/4C10-2.js","titre":"Signe d'un quotient de nombres relatifs"},"4P10-1":{"url":"/exercices/4e/4P10-1.js","titre":"Résoudre un problème de proportionnalité à l'aide d'un graphique"},"4G11":{"url":"/exercices/4e/4G11.js","titre":"Trouver l\'image d'une figure par une translation dans un pavage"},"4C33-3":{"url":"/exercices/4e/4C33-3.js","titre":"Puissances : Calculs automatisés et règles de calculs"},"4C33-1":{"url":"/exercices/4e/4C33-1.js","titre":"Puissances : Le sens des règles de calculs"},"4L14-2":{"url":"/exercices/4e/4L14-2.js","titre":"Tester si un nombre est solution d'une équation du second degré"},"4L10-1":{"url":"/exercices/4e/4L10-1.js","titre":"Réduire, si possible, une expression littérale simple"},"4L20":{"url":"/exercices/4e/4L20.js","titre":"Équation du premier degré"},"4C25-0":{"url":"/exercices/4e/4C25-0.js","titre":"Problèmes additifs et de comparaison sur les rationnels"},"4G31":{"url":"/exercices/4e/4G31.js","titre":"Contrôler si deux droites sont parallèles"},"4C11":{"url":"/exercices/4e/4C11.js","titre":"Calculs utilisant les priorités opératoires"},"4G40":{"url":"/exercices/4e/4G40.js","titre":"Utiliser le cosinus pour calculer une longueur dans un triangle rectangle"},"4C22-1":{"url":"/exercices/4e/4C22-1.js","titre":"Trouver l'inverse d'un nombre"},"4C32-1":{"url":"/exercices/4e/4C32-1.js","titre":"Calcul avec les puissances de dix"},"4G30-1":{"url":"/exercices/4e/4G30-1.js","titre":"Calculer des longueurs avec la propriété de Thalès (MG32)"},"4F12":{"url":"/exercices/4e/4F12.js","titre":"Problème s'appuyant sur la lecture d'une représentation graphique"},"4P10-2":{"url":"/exercices/4e/4P10-2.js","titre":"Déterminer une quatrième proportionnelle dans un tableau"},"4G22":{"url":"/exercices/4e/4G22.js","titre":"Problèmes utilisant le théorème de Pythagore"},"4C23":{"url":"/exercices/4e/4C23.js","titre":"Fractions et priorités opératoires"},"4C32":{"url":"/exercices/4e/4C32.js","titre":"Notation scientifique"},"4L10":{"url":"/exercices/4e/4L10.js","titre":"Utiliser la simple distributivité"},"4C30-3":{"url":"/exercices/4e/4C30-3.js","titre":"Écriture décimale d'une puissance"},"4L14-0":{"url":"/exercices/4e/4L14-0.js","titre":"Tester si un nombre est solution d'une équation"},"4L13-0":{"url":"/exercices/4e/4L13-0.js","titre":"Mettre en équation un problème sans objectif de résolution"},"4C30-1":{"url":"/exercices/4e/4C30-1.js","titre":"Encadrer avec des puissances de 10"},"4G10":{"url":"/exercices/4e/4G10.js","titre":"Trouver l'image d'un point par une symétrie axiale ou centrale ou par une translation"},"4G20-2":{"url":"/exercices/4e/4G20-2.js","titre":"Racine carré d'un carré parfait (calcul mental)"},"4G21":{"url":"/exercices/4e/4G21.js","titre":"Déterminer si un triangle est rectangle ou pas."},"4C21":{"url":"/exercices/4e/4C21.js","titre":"Additionner ou soustraire deux fractions"},"4C30-2":{"url":"/exercices/4e/4C30-2.js","titre":"Écriture décimale d'une puissance de 10"},"4L14-1":{"url":"/exercices/4e/4L14-1.js","titre":"Tester si un nombre est solution d'une équation du premier degré"},"4C10-1":{"url":"/exercices/4e/4C10-1.js","titre":"Signe d'un produit de nombres relatifs"},"4G30":{"url":"/exercices/4e/4G30.js","titre":"Calculer des longueurs avec la propriété de Thalès"},"4G20MG32":{"url":"/exercices/4e/4G20MG32.js","titre":"Calculer une longueur avec l'égalité de Pythagore (MG32)"},"4C10-5":{"url":"/exercices/4e/4C10-5.js","titre":"Multiplications de deux entiers relatifs dans un tableau à double entrée"},"4G20-1":{"url":"/exercices/4e/4G20-1.js","titre":"Donner ou compléter une égalité de Pythagore"},"4C10-3":{"url":"/exercices/4e/4C10-3.js","titre":"Multiplication de deux entiers relatifs"},"4G11-1":{"url":"/exercices/4e/4G11-1.js","titre":"Trouver l'image d'une figure par une translation dans un pavage (sortie Latex impossible)"},"4C21-1":{"url":"/exercices/4e/4C21-1.js","titre":"Additionner deux fractions"},"4C22":{"url":"/exercices/4e/4C22.js","titre":"Mutliplier des fractions"},"4G20":{"url":"/exercices/4e/4G20.js","titre":"Calculer une longueur avec le théorème de Pythagore"},"4G41":{"url":"/exercices/4e/4G41.js","titre":"Utiliser le cosinus pour calculer la mesure d'un angle dans un triangle rectangle"},"4G51":{"url":"/exercices/4e/4G51.js","titre":"Représenter un solide"},"4C22-2":{"url":"/exercices/4e/4C22-2.js","titre":"Diviser des fractions"},"4C10-0":{"url":"/exercices/4e/4C10-0.js","titre":"Signe d'un produit ou d'un quotient de nombres relatifs"},"4L13-1":{"url":"/exercices/4e/4L13-1.js","titre":"Produire une forme littérale en introduisant une lettre pour désigner une valeur inconnue"},"4Algo1-0":{"url":"/exercices/4e/4Algo1-0.js","titre":"Dessiner avec scratch"},"4C10-4":{"url":"/exercices/4e/4C10-4.js","titre":"Quotient de deux entiers relatifs"},"4L15-1":{"url":"/exercices/4e/4L15-1.js","titre":"Equations du type $\\dfrac{x}{a}=\\dfrac{b}{c}$"},"4P10":{"url":"/exercices/4e/4P10.js","titre":"Résoudre des problèmes de grandeurs composées et de conversion d'unités complexes"},"4L15-0":{"url":"/exercices/4e/4L15-0.js","titre":"Trouver l'erreur dans une résolution d'équation du premier degré"},"1E10":{"url":"/exercices/1e/1E10.js","titre":"Calcul du discriminant d'une équation du second degré"},"1E12":{"url":"/exercices/1e/1E12.js","titre":"Trouver l'équation d'une parabole"},"1E11":{"url":"/exercices/1e/1E11.js","titre":"Résoudre une équation du second degré"},"1N11":{"url":"/exercices/1e/1N11.js","titre":"Déterminer les termes d'une suite définie par récurrence"},"1N10":{"url":"/exercices/1e/1N10.js","titre":"Déterminer les termes d'une suite définie de façon explicite"},"1F10":{"url":"/exercices/1e/1F10.js","titre":"Calculs de dérivés"},"PEA11-1":{"url":"/exercices/PE/PEA11-1.js","titre":"Passer de la base 12 ou 16 à la base 10 et inversement"},"PEA11":{"url":"/exercices/PE/PEA11.js","titre":"Passer de la base 10 à une autre base et inversement"},"P004":{"url":"/exercices/Profs/P004.js","titre":"Droites graduées avec zoom"},"P007":{"url":"/exercices/Profs/P007.js","titre":"Fabriquer des pavages pour travailler les transformations"},"P003":{"url":"/exercices/Profs/P003.js","titre":"Tracer des droites graduées"},"P005":{"url":"/exercices/Profs/P005.js","titre":"Grilles décimales"},"P006":{"url":"/exercices/Profs/P006.js","titre":"Placer un nombre décimal avec des zooms successifs"},"6C23":{"url":"/exercices/6e/6C23.js","titre":"Reconnaitre un quadrilatère particulier à partir de ses propriétés"},"6N20-1":{"url":"/exercices/6e/6N20-1.js","titre":"Encadrer une fraction entre deux nombres entiers"},"6N11":{"url":"/exercices/6e/6N11.js","titre":"Lire l'abscisse entière d'un point (grands nombres)"},"6D101":{"url":"/exercices/6e/6D101.js","titre":"Utiliser les heures décimales"},"6N30-1":{"url":"/exercices/6e/6N30-1.js","titre":"Lire l'abscisse décimale d'un point repéré par une fraction"},"6G25-3":{"url":"/exercices/6e/6G25-3.js","titre":"Trouver l'image d'une figure par une symétrie axiale dans un pavage"},"6M30":{"url":"/exercices/6e/6M30.js","titre":"Calculs de volumes"},"6G24-3":{"url":"/exercices/6e/6G24-3.js","titre":"Construire le symétrique d'un point avec cible auto-corrective"},"6G42":{"url":"/exercices/6e/6G42.js","titre":"Connaître les propriétés du cube et du pavé droit"},"CM000":{"url":"/exercices/6e/CM000.js","titre":"Additions et de soustractions"},"6G23-2":{"url":"/exercices/6e/6G23-2.js","titre":"Tracer un triangle dont on connait une longueur et 2 angles"},"6C11":{"url":"/exercices/6e/6C11.js","titre":"Divisions euclidiennes"},"6C21":{"url":"/exercices/6e/6C21.js","titre":"Divisions euclidiennes - Niveau 2"},"6N11-2":{"url":"/exercices/6e/6N11-2.js","titre":"Placer un point d'abscisse entière (grands nombres)"},"6D11":{"url":"/exercices/6e/6D11.js","titre":"Additionner des durées"},"6N30":{"url":"/exercices/6e/6N30.js","titre":"Lire l'abscisse décimale d'un point"},"6G25":{"url":"/exercices/6e/6G25.js","titre":"Construire des médiatrices avec cible auto-corrective"},"6G33":{"url":"/exercices/6e/6G33.js","titre":"Reconnaitre un quadrilatère particulier à partir de ses propriétés"},"6Algo10":{"url":"/exercices/6e/6Algo10.js","titre":"Programmer des déplacements"},"6N21":{"url":"/exercices/6e/6N21.js","titre":"Lire l'abscisse fractionnaire d'un point"},"6C10-1":{"url":"/exercices/6e/6C10-1.js","titre":"Tables de multiplications"},"6N14":{"url":"/exercices/6e/6N14.js","titre":"Représenter des fractions"},"6G10":{"url":"/exercices/6e/6G10.js","titre":"Notation des droites, segments et demi-droites"},"6M20":{"url":"/exercices/6e/6M20.js","titre":"Aires de triangles"},"6G10-1":{"url":"/exercices/6e/6G10-1.js","titre":"Description et notation des droites, segments et demi-droites"},"6C10-4":{"url":"/exercices/6e/6C10-4.js","titre":"Addition de deux entiers"},"6test2":{"url":"/exercices/6e/6test2.js","titre":"Mon test"},"6G32-1":{"url":"/exercices/6e/6G32-1.js","titre":"Propriétés de conservation de la symétrie axiale"},"6G10-2":{"url":"/exercices/6e/6G10-2.js","titre":"Utiliser le codage pour décrire"},"6C13":{"url":"/exercices/6e/6C13.js","titre":"Traduire des phrases en calculs et réciproquement"},"6G43":{"url":"/exercices/6e/6G43.js","titre":"Connaître les propriétés du cube et du pavé droit"},"6C30":{"url":"/exercices/6e/6C30.js","titre":"Multiplications posées de nombres décimaux"},"6C10-2":{"url":"/exercices/6e/6C10-2.js","titre":"Tables de multiplications et multiples de 10"},"6G11":{"url":"/exercices/6e/6G11.js","titre":"Tracer des perpendiculaires"},"6C31-2":{"url":"/exercices/6e/6C31-2.js","titre":"Donner des valeurs approchées d'un quotient décimal"},"6N23-2":{"url":"/exercices/6e/6N23-2.js","titre":"Lire des abscisses décimales sous trois formes"},"6G41":{"url":"/exercices/6e/6G41.js","titre":"Compléter une représentation en perspective cavalière"},"6C10-5":{"url":"/exercices/6e/6C10-5.js","titre":"Labyrinthe de multiples"},"6G20-2":{"url":"/exercices/6e/6G20-2.js","titre":"Vocabulaire des triangles"},"6C32":{"url":"/exercices/6e/6C32.js","titre":"Problème - Les courses"},"6G24":{"url":"/exercices/6e/6G24.js","titre":"Trouver l'image d'un point par une symétrie axiale"},"6G12-1":{"url":"/exercices/6e/6G12-1.js","titre":"Tracer des parallèles et des perpendiculaires"},"6N23-3":{"url":"/exercices/6e/6N23-3.js","titre":"Lire une abscisse décimale grâce à des zooms successifs"},"6G25-1":{"url":"/exercices/6e/6G25-1.js","titre":"Trouver l'image d'une figure par une symétrie axiale dans un pavage triangulaire (sortie Latex Impossible)"},"6C10-3":{"url":"/exercices/6e/6C10-3.js","titre":"Tables de multiplications et nombres décimaux"},"6C33":{"url":"/exercices/6e/6C33.js","titre":"Calculer en utilisant les priorités opératoires"},"6C31":{"url":"/exercices/6e/6C31.js","titre":"Divisions décimales"},"6D12":{"url":"/exercices/6e/6D12.js","titre":"Calculer des durées ou déterminer un horaire"},"6D10":{"url":"/exercices/6e/6D10.js","titre":"Convertir des durées"},"6G25-2":{"url":"/exercices/6e/6G25-2.js","titre":"Trouver l'image d'une figure par une symétrie axiale dans un pavage carré (sortie Latex Impossible)"},"6G14":{"url":"/exercices/6e/6G14.js","titre":"Utiliser les propriétés des droites perpendiculaires"},"6test1":{"url":"/exercices/6e/6test1.js","titre":"Mon test"},"6CM01":{"url":"/exercices/6e/6CM01.js","titre":"Tables de multiplications"},"6C30-1":{"url":"/exercices/6e/6C30-1.js","titre":"Multiplications d'un nombre décimal par 10, 100 ou 1 000."},"6G20":{"url":"/exercices/6e/6G20.js","titre":"Nommer et coder des polygones"},"6G21":{"url":"/exercices/6e/6G21.js","titre":"Construire un triangle aux instruments"},"6C20":{"url":"/exercices/6e/6C20.js","titre":"Additions et soustractions de nombres décimaux"},"6G12":{"url":"/exercices/6e/6G12.js","titre":"Tracer des parallèles"},"6C30-2":{"url":"/exercices/6e/6C30-2.js","titre":"Calculer le produit de deux décimaux connaissant le produit de deux entiers"},"6N30-2":{"url":"/exercices/6e/6N30-2.js","titre":"Placer un point d'abscisse décimale"},"5C11":{"url":"/exercices/5e/5C11.js","titre":"Traduire une phrase par une expression"},"5R22":{"url":"/exercices/5e/5R22.js","titre":"Additions et soustractions de nombres relatifs"},"5R10-0":{"url":"/exercices/5e/5R10-0.js","titre":"Trouver l'opposé d'un nombre relatif"},"5L13-1":{"url":"/exercices/5e/5L13-1.js","titre":"Calculer une expression littérale pour les valeurs données en détaillant les calculs"},"5R20-3":{"url":"/exercices/5e/5R20-3.js","titre":"Additions de 5 nombres relatifs"},"5C12":{"url":"/exercices/5e/5C12.js","titre":"Calculer une expression numérique en détaillant les calculs"},"5G11-1":{"url":"/exercices/5e/5G11-1.js","titre":"Construire par symétrie axiale"},"5L15":{"url":"/exercices/5e/5L15.js","titre":"Tester une égalité"},"5G10":{"url":"/exercices/5e/5G10.js","titre":"Trouver l'image d'un point par une symétrie axiale"},"5R11-2":{"url":"/exercices/5e/5R11-2.js","titre":"Placer un point d'abscisse un nombre relatif"},"5D10":{"url":"/exercices/5e/5D10.js","titre":"Convertir des durées"},"5C12-1":{"url":"/exercices/5e/5C12-1.js","titre":"Traduire une phrase par une expression et la calculer"},"5S14":{"url":"/exercices/5e/5S14.js","titre":"Calculer des moyennes"},"5M10":{"url":"/exercices/5e/5M10.js","titre":"Aire du parallélogramme"},"5C10":{"url":"/exercices/5e/5C10.js","titre":"Ajouter 9"},"5N14":{"url":"/exercices/5e/5N14.js","titre":"Comparer des fractions (dénominateurs multiples)"},"5G30":{"url":"/exercices/5e/5G30.js","titre":"Utiliser le codage pour décrire"},"5L10-3":{"url":"/exercices/5e/5L10-3.js","titre":"Traduire une expression par une phrase"},"5N20":{"url":"/exercices/5e/5N20.js","titre":"Additionner ou soustraire deux fractions (dénominateurs multiples)"},"5G11-2":{"url":"/exercices/5e/5G11-2.js","titre":"Construire par symétrie centrale"},"5A11-1":{"url":"/exercices/5e/5A11-1.js","titre":"Labyrinthe de multiples basé sur les critères de divisibilité"},"5L12":{"url":"/exercices/5e/5L12.js","titre":"Réduire une expression littérale"},"5L14-4":{"url":"/exercices/5e/5L14-4.js","titre":"Déterminer la dernière opération à effectuer dans une expression littérale"},"5L13-3":{"url":"/exercices/5e/5L13-3.js","titre":"Traduire une phrase par une expression et la calculer"},"5L12-1":{"url":"/exercices/5e/5L12-1.js","titre":"Réduire un produit et une somme à partir des mêmes éléments algébriques pour distinguer la différence"},"5N14-2":{"url":"/exercices/5e/5N14-2.js","titre":"Comparer quatre fractions (dénominateurs multiples) et un nombre entier"},"5R12-2":{"url":"/exercices/5e/5R12-2.js","titre":"Déterminer les coordonnées (relatives) d'un point"},"5L10-4":{"url":"/exercices/5e/5L10-4.js","titre":"Produire une formule à partir d'un tableau"},"5G11-3":{"url":"/exercices/5e/5G11-3.js","titre":"Construire le symétrique d'un point avec cible auto-corrective"},"5A10":{"url":"/exercices/5e/5A10.js","titre":"Écrire la liste de tous les diviseurs d'un entier."},"5A12-1":{"url":"/exercices/5e/5A12-1.js","titre":"Primalité ou pas"},"5S13":{"url":"/exercices/5e/5S13.js","titre":"Calculer des fréquences"},"5M20":{"url":"/exercices/5e/5M20.js","titre":"Calculs de volumes"},"5G11":{"url":"/exercices/5e/5G11.js","titre":"Trouver l'image d'un point par une symétrie axiale ou centrale"},"5N11-1":{"url":"/exercices/5e/5N11-1.js","titre":"Tableaux et pourcentages - prix constant"},"5G12-1":{"url":"/exercices/5e/5G12-1.js","titre":"Trouver l'image d'une figure par une symétrie centrale dans un pavage (sortie Latex Impossible)"},"5P10":{"url":"/exercices/5e/5P10.js","titre":"Tableaux et proportionnalité."},"5G20":{"url":"/exercices/5e/5G20.js","titre":"Construire un triangle avec cible auto-corrective"},"5G20-1":{"url":"/exercices/5e/5G20-1.js","titre":"Vocabulaire des triangles"},"5G21-1":{"url":"/exercices/5e/5G21-1.js","titre":"Constructibilité des triangles via les longueurs"},"5L10":{"url":"/exercices/5e/5L10.js","titre":"Écrire une expression littérale"},"5L10-2":{"url":"/exercices/5e/5L10-2.js","titre":"Traduire un programme de calcul par une expression littérale"},"5R11":{"url":"/exercices/5e/5R11.js","titre":"Lire l'abscisse relative d'un point"},"5R20-4":{"url":"/exercices/5e/5R20-4.js","titre":"Résoudre un problème en utilisant une somme algébrique de relatifs."},"5R20-5":{"url":"/exercices/5e/5R20-5.js","titre":"Additions de deux entiers relatifs dans un tableau à double entrée"},"5R20":{"url":"/exercices/5e/5R20.js","titre":"Addition de deux entiers relatifs"},"5L10-1":{"url":"/exercices/5e/5L10-1.js","titre":"Traduire une phrase par une expression"},"5L13-2":{"url":"/exercices/5e/5L13-2.js","titre":"Substitution"},"5CM01":{"url":"/exercices/5e/5CM01.js","titre":"Tables de multiplications"},"5S20":{"url":"/exercices/5e/5S20.js","titre":"Placer un événement sur une échelle de probabilités"},"5A13":{"url":"/exercices/5e/5A13.js","titre":"Décomposition en facteurs premiers"},"5G12":{"url":"/exercices/5e/5G12.js","titre":"Symétrie centrale dans un pavage"},"5G31-1":{"url":"/exercices/5e/5G31-1.js","titre":"Constructibilité des triangles via les angles"},"5N110":{"url":"/exercices/5e/5N110.js","titre":"Variation en pourcentages"},"5L14":{"url":"/exercices/5e/5L14.js","titre":"Calculer la valeur d'une expression littérale"},"5L13":{"url":"/exercices/5e/5L13.js","titre":"Réduire une expression de la forme $ax+bx$"},"5S21":{"url":"/exercices/5e/5S21.js","titre":"Calculer des probabilités dans une expérience aléatoire à une épreuve"},"5R22-2":{"url":"/exercices/5e/5R22-2.js","titre":"Simplifier l'écriture d'une somme de 2 relatifs et calculer"},"5N11-2":{"url":"/exercices/5e/5N11-2.js","titre":"Tableaux et pourcentages - pourcentage constant"},"5R21":{"url":"/exercices/5e/5R21.js","titre":"Soustraction de deux entiers relatifs"},"5G22":{"url":"/exercices/5e/5G22.js","titre":"Déterminer la nature d'une droite remarquable"},"5C11-2":{"url":"/exercices/5e/5C11-2.js","titre":"Traduire une succession des opérations par une expression"},"5R20-2":{"url":"/exercices/5e/5R20-2.js","titre":"Addition à trou de deux entiers relatifs"},"5N20-0":{"url":"/exercices/5e/5N20-0.js","titre":"Résoudre un problème en utilisant des fractions"},"5G31":{"url":"/exercices/5e/5G31.js","titre":"Somme des angles dans un triangle"},"5G30-1":{"url":"/exercices/5e/5G30-1.js","titre":"Déterminer des angles en utilisant les cas d'égalité"},"5C11-1":{"url":"/exercices/5e/5C11-1.js","titre":"Traduire une expression par une phrase"},"3F12-2":{"url":"/exercices/3e/3F12-2.js","titre":"Déterminer l'image d'un nombre par une fonction d'après sa forme algébrique"},"3A13":{"url":"/exercices/3e/3A13.js","titre":"Engrenages"},"3G10-1":{"url":"/exercices/3e/3G10-1.js","titre":"Trouver les coordonnées de l'image d'un point par une transformation du plan"},"3F21-1":{"url":"/exercices/3e/3F21-1.js","titre":"Déterminer une fonction affine"},"3A11-1":{"url":"/exercices/3e/3A11-1.js","titre":"Primalité ou pas - Variante avec les critères de divisibilité par 7 et par 11"},"3P10-1":{"url":"/exercices/3e/3P10-1.js","titre":"Coefficient multiplicateur d'une variation en pourcentage"},"3G12-1":{"url":"/exercices/3e/3G12-1.js","titre":"Trouver l'image d'une figure par une rotation de 90 degrés dans un pavage (sortie Latex Impossible)"},"3A10":{"url":"/exercices/3e/3A10.js","titre":"Division Euclidienne - Diviseurs - Multiples"},"3A12":{"url":"/exercices/3e/3A12.js","titre":"Fractions irréductibles"},"3L14":{"url":"/exercices/3e/3L14.js","titre":"Résoudre une équation produit nul"},"3L13-2":{"url":"/exercices/3e/3L13-2.js","titre":"Equations résolvantes pour le théorème de Thalès"},"3G22":{"url":"/exercices/3e/3G22.js","titre":"Connaître les effets des agrandissements/réductions sur les aires et les volumes"},"3F12":{"url":"/exercices/3e/3F12.js","titre":"Fonctions : Calculs d'images"},"3G23":{"url":"/exercices/3e/3G23.js","titre":"Reconnaître des triangles semblables dans différentes configurations"},"3G30":{"url":"/exercices/3e/3G30.js","titre":"Déterminer une longueur grâce à la trigonométrie"},"3G10-2":{"url":"/exercices/3e/3G10-2.js","titre":"Trouver l'image d'un point par une transformation choisie aléatoirement"},"3G21":{"url":"/exercices/3e/3G21.js","titre":"Contrôler si deux droites sont parallèles"},"3Algo1":{"url":"/exercices/3e/3Algo1.js","titre":"Instruction conditionelle"},"3G10-3":{"url":"/exercices/3e/3G10-3.js","titre":"Construire l'image d'un point par une rotation avec cible auto-corrective"},"3S20":{"url":"/exercices/3e/3S20.js","titre":"Calculer des probabilités dans une expérience aléatoire à deux épreuves"},"3G20":{"url":"/exercices/3e/3G20.js","titre":"Calculer des longueurs avec la propriété de Thalès"},"3S15":{"url":"/exercices/3e/3S15.js","titre":"Calculer des étendues"},"3A11-2":{"url":"/exercices/3e/3A11-2.js","titre":"Décomposition en facteurs premiers d'un entier"},"3L11-1":{"url":"/exercices/3e/3L11-1.js","titre":"Utiliser la double distributivité"},"3F13-1":{"url":"/exercices/3e/3F13-1.js","titre":"Lecture graphique d'images et d'antécédents."},"3L12":{"url":"/exercices/3e/3L12.js","titre":"Factoriser a²-b²"},"3L10-1":{"url":"/exercices/3e/3L10-1.js","titre":"Additionner ou soustraire une expression entre parenthèses"},"3F10":{"url":"/exercices/3e/3F10.js","titre":"Lectures d'images et d'antécédents depuis un tableau de valeurs"},"3L13-1":{"url":"/exercices/3e/3L13-1.js","titre":"Équation du premier degré (utilisant la distributivité)"},"3L11":{"url":"/exercices/3e/3L11.js","titre":"Utiliser la simple distributivité"},"3G11":{"url":"/exercices/3e/3G11.js","titre":"Construire l'image d'un point par une homothetie avec cible auto-corrective"},"3G43":{"url":"/exercices/3e/3G43.js","titre":"Calculs de volumes"},"3F12-3":{"url":"/exercices/3e/3F12-3.js","titre":"Compléter un tableau de valeurs"},"3A11":{"url":"/exercices/3e/3A11.js","titre":"Primalité ou pas"},"3P10":{"url":"/exercices/3e/3P10.js","titre":"Variations en pourcentage"},"3G12":{"url":"/exercices/3e/3G12.js","titre":"Trouver l'image d'une figure par une rotation dans un pavage"},"3L11-4":{"url":"/exercices/3e/3L11-4.js","titre":"Factoriser une expression"},"3L15":{"url":"/exercices/3e/3L15.js","titre":"Résoudre une équation du second degré"},"3F1-act":{"url":"/exercices/3e/3F1-act.js","titre":"Fonctions : Notion et vocabulaire"},"3L10":{"url":"/exercices/3e/3L10.js","titre":"Donner l'opposé d'une expression"},"3F12-4":{"url":"/exercices/3e/3F12-4.js","titre":"Lire l'image d'un nombre à partir d'un graphique"},"3F13-2":{"url":"/exercices/3e/3F13-2.js","titre":"Spécial escape game"},"3L12-1":{"url":"/exercices/3e/3L12-1.js","titre":"Développer (a-b)(a+b)"},"3L11-3":{"url":"/exercices/3e/3L11-3.js","titre":"Utiliser la distributivité (simple ou double) et réduire"},"3A11-3":{"url":"/exercices/3e/3A11-3.js","titre":"Compter/lister les diviseurs d'un entier à partir de sa décomposition en facteurs premiers."},"3G20-1":{"url":"/exercices/3e/3G20-1.js","titre":"Problèmes avec le théorème de Thalès"},"3G31":{"url":"/exercices/3e/3G31.js","titre":"Déterminer un angle grâce à la trigonométrie"},"3L11-5":{"url":"/exercices/3e/3L11-5.js","titre":"Calcul mental et calcul littéral"},"3L14-1":{"url":"/exercices/3e/3L14-1.js","titre":"Résoudre une équation produit nul (niveau 2)"},"3G20-2":{"url":"/exercices/3e/3G20-2.js","titre":"Déterminer une longueur avec la propriété de Thales (MG32)"},"3L13":{"url":"/exercices/3e/3L13.js","titre":"Équation du premier degré"},"3F13":{"url":"/exercices/3e/3F13.js","titre":"Lire les antécédents d'un nombre à partir d'un graphique"},"3L11-2":{"url":"/exercices/3e/3L11-2.js","titre":"Réduire une expression"},"3F21":{"url":"/exercices/3e/3F21.js","titre":"Déterminer une fonction linéaire"},"2L11":{"url":"/exercices/2e/2L11.js","titre":"Factoriser avec les identités remarquables"},"2N20":{"url":"/exercices/2e/2N20.js","titre":"Déterminer le plus petit ensemble de nombres dans lequel le nombre proposé appartient"},"2G10":{"url":"/exercices/2e/2G10.js","titre":"Utiliser la distance entre deux points dans un repère orthonormé"},"2N10":{"url":"/exercices/2e/2N10.js","titre":"Existence d'une racine carrée"},"2N12":{"url":"/exercices/2e/2N12.js","titre":"Appliquer la double distributivité avec les racines carrées"},"2N21":{"url":"/exercices/2e/2N21.js","titre":"Déterminer la parité d'une expression"},"2N24":{"url":"/exercices/2e/2N24.js","titre":"Associer un intervalle de  $\\mathbb{R}$ à une inéquation et son schéma sur une droite graduée"},"2N25":{"url":"/exercices/2e/2N25.js","titre":"Utiliser et comprendre les symboles $\\cup $ et $\\cap $ avec les intervalles de $\\mathbb{R}$"},"2N12-1":{"url":"/exercices/2e/2N12-1.js","titre":"Développer les identités remarquables avec des racines carrées"},"2N11-1":{"url":"/exercices/2e/2N11-1.js","titre":"Simplifier une somme de racines carrées"},"2N22":{"url":"/exercices/2e/2N22.js","titre":"Utiliser la notion de valeur absolue d'une quantité"},"2N11":{"url":"/exercices/2e/2N11.js","titre":"Ecrire le nombre proposé sous la forme $a\\sqrt{b}$"},"2L10":{"url":"/exercices/2e/2L10.js","titre":"Développer avec les identités remarquables"},"2N10-1":{"url":"/exercices/2e/2N10-1.js","titre":"Connaître les propriétés calculatoires des racines carrées"},"2G11":{"url":"/exercices/2e/2G11.js","titre":"Déterminer les coordonnées milieu d’un segment dans un repère"},"2G12":{"url":"/exercices/2e/2G12.js","titre":"Déterminer la nature d'un polygone."},"2N23":{"url":"/exercices/2e/2N23.js","titre":"Résoudre une équation avec des valeurs absolues"},};
-
+export let dictionnaireDesExercices = {"4G53":{"url":"/exercices/4e/4G53.js","titre":"Calculs de volumes"},"4C30":{"url":"/exercices/4e/4C30.js","titre":"Puissances de 10 : Le sens des règles de calculs"},"4C10-2":{"url":"/exercices/4e/4C10-2.js","titre":"Signe d'un quotient de nombres relatifs"},"4P10-1":{"url":"/exercices/4e/4P10-1.js","titre":"Résoudre un problème de proportionnalité à l'aide d'un graphique"},"4G11":{"url":"/exercices/4e/4G11.js","titre":"Trouver l\'image d'une figure par une translation dans un pavage"},"4C33-3":{"url":"/exercices/4e/4C33-3.js","titre":"Puissances : Calculs automatisés et règles de calculs"},"4C33-1":{"url":"/exercices/4e/4C33-1.js","titre":"Puissances : Le sens des règles de calculs"},"4L14-2":{"url":"/exercices/4e/4L14-2.js","titre":"Tester si un nombre est solution d'une équation du second degré"},"4L10-1":{"url":"/exercices/4e/4L10-1.js","titre":"Réduire, si possible, une expression littérale simple"},"4L20":{"url":"/exercices/4e/4L20.js","titre":"Équation du premier degré"},"4C25-0":{"url":"/exercices/4e/4C25-0.js","titre":"Problèmes additifs et de comparaison sur les rationnels"},"4G31":{"url":"/exercices/4e/4G31.js","titre":"Contrôler si deux droites sont parallèles"},"4C11":{"url":"/exercices/4e/4C11.js","titre":"Calculs utilisant les priorités opératoires"},"4G40":{"url":"/exercices/4e/4G40.js","titre":"Utiliser le cosinus pour calculer une longueur dans un triangle rectangle"},"4C22-1":{"url":"/exercices/4e/4C22-1.js","titre":"Trouver l'inverse d'un nombre"},"4C32-1":{"url":"/exercices/4e/4C32-1.js","titre":"Calcul avec les puissances de dix"},"4G30-1":{"url":"/exercices/4e/4G30-1.js","titre":"Calculer des longueurs avec la propriété de Thalès (MG32)"},"4F12":{"url":"/exercices/4e/4F12.js","titre":"Problème s'appuyant sur la lecture d'une représentation graphique"},"4P10-2":{"url":"/exercices/4e/4P10-2.js","titre":"Déterminer une quatrième proportionnelle dans un tableau"},"4G22":{"url":"/exercices/4e/4G22.js","titre":"Problèmes utilisant le théorème de Pythagore"},"4C23":{"url":"/exercices/4e/4C23.js","titre":"Fractions et priorités opératoires"},"4C32":{"url":"/exercices/4e/4C32.js","titre":"Notation scientifique"},"4L10":{"url":"/exercices/4e/4L10.js","titre":"Utiliser la simple distributivité"},"4C30-3":{"url":"/exercices/4e/4C30-3.js","titre":"Écriture décimale d'une puissance"},"4L14-0":{"url":"/exercices/4e/4L14-0.js","titre":"Tester si un nombre est solution d'une équation"},"4L13-0":{"url":"/exercices/4e/4L13-0.js","titre":"Mettre en équation un problème sans objectif de résolution"},"4C30-1":{"url":"/exercices/4e/4C30-1.js","titre":"Encadrer avec des puissances de 10"},"4G10":{"url":"/exercices/4e/4G10.js","titre":"Trouver l'image d'un point par une symétrie axiale ou centrale ou par une translation"},"4G20-2":{"url":"/exercices/4e/4G20-2.js","titre":"Racine carré d'un carré parfait (calcul mental)"},
+"4G21":{"url":"/exercices/4e/4G21.js","titre":"Déterminer si un triangle est rectangle ou pas."},"4C21":{"url":"/exercices/4e/4C21.js","titre":"Additionner ou soustraire deux fractions"},"4C30-2":{"url":"/exercices/4e/4C30-2.js","titre":"Écriture décimale d'une puissance de 10"},"4L14-1":{"url":"/exercices/4e/4L14-1.js","titre":"Tester si un nombre est solution d'une équation du premier degré"},"4C10-1":{"url":"/exercices/4e/4C10-1.js","titre":"Signe d'un produit de nombres relatifs"},"4G30":{"url":"/exercices/4e/4G30.js","titre":"Calculer des longueurs avec la propriété de Thalès"},"4G20MG32":{"url":"/exercices/4e/4G20MG32.js","titre":"Calculer une longueur avec l'égalité de Pythagore (MG32)"},"4C10-5":{"url":"/exercices/4e/4C10-5.js","titre":"Multiplications de deux entiers relatifs dans un tableau à double entrée"},"4G20-1":{"url":"/exercices/4e/4G20-1.js","titre":"Donner ou compléter une égalité de Pythagore"},"4C10-3":{"url":"/exercices/4e/4C10-3.js","titre":"Multiplication de deux entiers relatifs"},"4G11-1":{"url":"/exercices/4e/4G11-1.js","titre":"Trouver l'image d'une figure par une translation dans un pavage (sortie Latex impossible)"},"4C21-1":{"url":"/exercices/4e/4C21-1.js","titre":"Additionner deux fractions"},"4C22":{"url":"/exercices/4e/4C22.js","titre":"Mutliplier des fractions"},"4G20":{"url":"/exercices/4e/4G20.js","titre":"Calculer une longueur avec le théorème de Pythagore"},"4G41":{"url":"/exercices/4e/4G41.js","titre":"Utiliser le cosinus pour calculer la mesure d'un angle dans un triangle rectangle"},"4G51":{"url":"/exercices/4e/4G51.js","titre":"Représenter un solide"},"4C22-2":{"url":"/exercices/4e/4C22-2.js","titre":"Diviser des fractions"},"4C10-0":{"url":"/exercices/4e/4C10-0.js","titre":"Signe d'un produit ou d'un quotient de nombres relatifs"},"4L13-1":{"url":"/exercices/4e/4L13-1.js","titre":"Produire une forme littérale en introduisant une lettre pour désigner une valeur inconnue"},"4Algo1-0":{"url":"/exercices/4e/4Algo1-0.js","titre":"Dessiner avec scratch"},"4C10-4":{"url":"/exercices/4e/4C10-4.js","titre":"Quotient de deux entiers relatifs"},"4L15-1":{"url":"/exercices/4e/4L15-1.js","titre":"Equations du type $\\dfrac{x}{a}=\\dfrac{b}{c}$"},"4P10":{"url":"/exercices/4e/4P10.js","titre":"Résoudre des problèmes de grandeurs composées et de conversion d'unités complexes"},"4L15-0":{"url":"/exercices/4e/4L15-0.js","titre":"Trouver l'erreur dans une résolution d'équation du premier degré"},"1E10":{"url":"/exercices/1e/1E10.js","titre":"Calcul du discriminant d'une équation du second degré"},"1E12":{"url":"/exercices/1e/1E12.js","titre":"Trouver l'équation d'une parabole"},"1E11":{"url":"/exercices/1e/1E11.js","titre":"Résoudre une équation du second degré"},"1N11":{"url":"/exercices/1e/1N11.js","titre":"Déterminer les termes d'une suite définie par récurrence"},"1N10":{"url":"/exercices/1e/1N10.js","titre":"Déterminer les termes d'une suite définie de façon explicite"},"1F10":{"url":"/exercices/1e/1F10.js","titre":"Calculs de dérivés"},"PEA11-1":{"url":"/exercices/PE/PEA11-1.js","titre":"Passer de la base 12 ou 16 à la base 10 et inversement"},"PEA11":{"url":"/exercices/PE/PEA11.js","titre":"Passer de la base 10 à une autre base et inversement"},"P004":{"url":"/exercices/Profs/P004.js","titre":"Droites graduées avec zoom"},"P007":{"url":"/exercices/Profs/P007.js","titre":"Fabriquer des pavages pour travailler les transformations"},"P003":{"url":"/exercices/Profs/P003.js","titre":"Tracer des droites graduées"},"P005":{"url":"/exercices/Profs/P005.js","titre":"Grilles décimales"},"P006":{"url":"/exercices/Profs/P006.js","titre":"Placer un nombre décimal avec des zooms successifs"},"6C23":{"url":"/exercices/6e/6C23.js","titre":"Reconnaitre un quadrilatère particulier à partir de ses propriétés"},"6N20-1":{"url":"/exercices/6e/6N20-1.js","titre":"Encadrer une fraction entre deux nombres entiers"},"6N11":{"url":"/exercices/6e/6N11.js","titre":"Lire l'abscisse entière d'un point (grands nombres)"},
+"6D101":{"url":"/exercices/6e/6D101.js","titre":"Utiliser les heures décimales"},"6N30-1":{"url":"/exercices/6e/6N30-1.js","titre":"Lire l'abscisse décimale d'un point repéré par une fraction"},"6G25-3":{"url":"/exercices/6e/6G25-3.js","titre":"Trouver l'image d'une figure par une symétrie axiale dans un pavage"},"6M30":{"url":"/exercices/6e/6M30.js","titre":"Calculs de volumes"},"6G24-3":{"url":"/exercices/6e/6G24-3.js","titre":"Construire le symétrique d'un point avec cible auto-corrective"},"6G42":{"url":"/exercices/6e/6G42.js","titre":"Connaître les propriétés du cube et du pavé droit"},"CM000":{"url":"/exercices/6e/CM000.js","titre":"Additions et de soustractions"},"6G23-2":{"url":"/exercices/6e/6G23-2.js","titre":"Tracer un triangle dont on connait une longueur et 2 angles"},"6C11":{"url":"/exercices/6e/6C11.js","titre":"Divisions euclidiennes"},"6C21":{"url":"/exercices/6e/6C21.js","titre":"Divisions euclidiennes - Niveau 2"},"6N11-2":{"url":"/exercices/6e/6N11-2.js","titre":"Placer un point d'abscisse entière (grands nombres)"},"6D11":{"url":"/exercices/6e/6D11.js","titre":"Additionner des durées"},"6N30":{"url":"/exercices/6e/6N30.js","titre":"Lire l'abscisse décimale d'un point"},"6G25":{"url":"/exercices/6e/6G25.js","titre":"Construire des médiatrices avec cible auto-corrective"},"6G33":{"url":"/exercices/6e/6G33.js","titre":"Reconnaitre un quadrilatère particulier à partir de ses propriétés"},"6G21-1":{"url":"/exercices/6e/6G21-1.js","titre":"Construire un triangle avec cible"},"6Algo10":{"url":"/exercices/6e/6Algo10.js","titre":"Programmer des déplacements"},"6N21":{"url":"/exercices/6e/6N21.js","titre":"Lire l'abscisse fractionnaire d'un point"},"6C10-1":{"url":"/exercices/6e/6C10-1.js","titre":"Tables de multiplication"},"6N14":{"url":"/exercices/6e/6N14.js","titre":"Représenter des fractions"},"6G10":{"url":"/exercices/6e/6G10.js","titre":"Notation des droites, segments et demi-droites"},"6M20":{"url":"/exercices/6e/6M20.js","titre":"Aires de triangles"},"6G10-1":{"url":"/exercices/6e/6G10-1.js","titre":"Description et notation des droites, segments et demi-droites"},"6C10-4":{"url":"/exercices/6e/6C10-4.js","titre":"Addition de deux entiers"},"6test2":{"url":"/exercices/6e/6test2.js","titre":"Mon test"},"6G32-1":{"url":"/exercices/6e/6G32-1.js","titre":"Propriétés de conservation de la symétrie axiale"},"6G10-2":{"url":"/exercices/6e/6G10-2.js","titre":"Utiliser le codage pour décrire"},"6C13":{"url":"/exercices/6e/6C13.js","titre":"Traduire des phrases en calculs et réciproquement"},"6G11-2":{"url":"/exercices/6e/6G11-2.js","titre":"Construire le symétrique d'une figure par rapport à un point"},"6G43":{"url":"/exercices/6e/6G43.js","titre":"Connaître les propriétés du cube et du pavé droit"},"6C30":{"url":"/exercices/6e/6C30.js","titre":"Multiplications posées de nombres décimaux"},"6C10-2":{"url":"/exercices/6e/6C10-2.js","titre":"Tables de multiplications et multiples de 10"},"6G11":{"url":"/exercices/6e/6G11.js","titre":"Tracer des perpendiculaires"},"6C31-2":{"url":"/exercices/6e/6C31-2.js","titre":"Donner des valeurs approchées d'un quotient décimal"},"6G11-1":{"url":"/exercices/6e/6G11-1.js","titre":"Construire le symétrique d'un point par rapport à une droite"},"6N23-2":{"url":"/exercices/6e/6N23-2.js","titre":"Lire des abscisses décimales sous trois formes"},"6G41":{"url":"/exercices/6e/6G41.js","titre":"Compléter une représentation en perspective cavalière"},"6C10-5":{"url":"/exercices/6e/6C10-5.js","titre":"Labyrinthe de multiples"},"6G20-2":{"url":"/exercices/6e/6G20-2.js","titre":"Vocabulaire des triangles"},"6C32":{"url":"/exercices/6e/6C32.js","titre":"Problème - Les courses"},"6G24":{"url":"/exercices/6e/6G24.js","titre":"Trouver l'image d'un point par une symétrie axiale"},"6G12-1":{"url":"/exercices/6e/6G12-1.js","titre":"Tracer des parallèles et des perpendiculaires"},"6N23-3":{"url":"/exercices/6e/6N23-3.js","titre":"Lire une abscisse décimale grâce à des zooms successifs"},"6G25-1":{"url":"/exercices/6e/6G25-1.js","titre":"Trouver l'image d'une figure par une symétrie axiale dans un pavage triangulaire (sortie Latex Impossible)"},"6G24-2":{"url":"/exercices/6e/6G24-2.js","titre":"Construire le symétrique d'une figure par rapport à une droite (cas simples)"},"6C10-3":{"url":"/exercices/6e/6C10-3.js","titre":"Tables de multiplications et nombres décimaux"},"6C33":{"url":"/exercices/6e/6C33.js","titre":"Calculer en utilisant les priorités opératoires"},"6C31":{"url":"/exercices/6e/6C31.js","titre":"Divisions décimales"},"6D12":{"url":"/exercices/6e/6D12.js","titre":"Calculer des durées ou déterminer un horaire"},"6G24-1":{"url":"/exercices/6e/6G24-1.js","titre":"Construire le symétrique d'un point par rapport à une droite (cas simples)"},"6D10":{"url":"/exercices/6e/6D10.js","titre":"Convertir des durées"},"6G25-2":{"url":"/exercices/6e/6G25-2.js","titre":"Trouver l'image d'une figure par une symétrie axiale dans un pavage carré (sortie Latex Impossible)"},"6G14":{"url":"/exercices/6e/6G14.js","titre":"Utiliser les propriétés des droites perpendiculaires"},"6C30-1":{"url":"/exercices/6e/6C30-1.js","titre":"Multiplications d'un nombre décimal par 10, 100 ou 1 000."},"6G20":{"url":"/exercices/6e/6G20.js","titre":"Nommer et coder des polygones"},"6C101":{"url":"/exercices/6e/6C101.js","titre":"Tables de multiplications"},"6G21":{"url":"/exercices/6e/6G21.js","titre":"Construire un triangle aux instruments"},"6C20":{"url":"/exercices/6e/6C20.js","titre":"Additions et soustractions de nombres décimaux"},"6G12":{"url":"/exercices/6e/6G12.js","titre":"Tracer des parallèles"},"6C30-2":{"url":"/exercices/6e/6C30-2.js","titre":"Calculer le produit de deux décimaux connaissant le produit de deux entiers"},"6N30-2":{"url":"/exercices/6e/6N30-2.js","titre":"Placer un point d'abscisse décimale"},"5C11":{"url":"/exercices/5e/5C11.js","titre":"Traduire une phrase par une expression"},"5R22":{"url":"/exercices/5e/5R22.js","titre":"Additions et soustractions de nombres relatifs"},"5R10-0":{"url":"/exercices/5e/5R10-0.js","titre":"Trouver l'opposé d'un nombre relatif"},"5L13-1":{"url":"/exercices/5e/5L13-1.js","titre":"Calculer une expression littérale pour les valeurs données en détaillant les calculs"},"5R20-3":{"url":"/exercices/5e/5R20-3.js","titre":"Additions de 5 nombres relatifs"},"5C12":{"url":"/exercices/5e/5C12.js","titre":"Calculer une expression numérique en détaillant les calculs"},"5G11-1":{"url":"/exercices/5e/5G11-1.js","titre":"Construire par symétrie axiale"},"5L15":{"url":"/exercices/5e/5L15.js","titre":"Tester une égalité"},"5G10":{"url":"/exercices/5e/5G10.js","titre":"Trouver l'image d'un point par une symétrie axiale"},"5R11-2":{"url":"/exercices/5e/5R11-2.js","titre":"Placer un point d'abscisse un nombre relatif"},"5D10":{"url":"/exercices/5e/5D10.js","titre":"Convertir des durées"},"5C12-1":{"url":"/exercices/5e/5C12-1.js","titre":"Traduire une phrase par une expression et la calculer"},"5S14":{"url":"/exercices/5e/5S14.js","titre":"Calculer des moyennes"},"5M10":{"url":"/exercices/5e/5M10.js","titre":"Aire du parallélogramme"},"5C10":{"url":"/exercices/5e/5C10.js","titre":"Ajouter 9"},"5N14":{"url":"/exercices/5e/5N14.js","titre":"Comparer des fractions (dénominateurs multiples)"},"5G30":{"url":"/exercices/5e/5G30.js","titre":"Utiliser le codage pour décrire"},"5L10-3":{"url":"/exercices/5e/5L10-3.js","titre":"Traduire une expression par une phrase"},"5N20":{"url":"/exercices/5e/5N20.js","titre":"Additionner ou soustraire deux fractions (dénominateurs multiples)"},"5G11-2":{"url":"/exercices/5e/5G11-2.js","titre":"Construire par symétrie centrale"},"5A11-1":{"url":"/exercices/5e/5A11-1.js","titre":"Labyrinthe de multiples basé sur les critères de divisibilité"},"5L12":{"url":"/exercices/5e/5L12.js","titre":"Réduire une expression littérale"},"5G10-2":{"url":"/exercices/5e/5G10-2.js","titre":"Construire le symétrique d'une figure par rapport à une droite"},"5L14-4":{"url":"/exercices/5e/5L14-4.js","titre":"Déterminer la dernière opération à effectuer dans une expression littérale"},"5L13-3":{"url":"/exercices/5e/5L13-3.js","titre":"Traduire une phrase par une expression et la calculer"},"5L12-1":{"url":"/exercices/5e/5L12-1.js","titre":"Réduire un produit et une somme à partir des mêmes éléments algébriques pour distinguer la différence"},"5N14-2":{"url":"/exercices/5e/5N14-2.js","titre":"Comparer quatre fractions (dénominateurs multiples) et un nombre entier"},"5R12-2":{"url":"/exercices/5e/5R12-2.js","titre":"Déterminer les coordonnées (relatives) d'un point"},"5L10-4":{"url":"/exercices/5e/5L10-4.js","titre":"Produire une formule à partir d'un tableau"},"5G11-3":{"url":"/exercices/5e/5G11-3.js","titre":"Construire le symétrique d'un point avec cible auto-corrective"},"5A10":{"url":"/exercices/5e/5A10.js","titre":"Écrire la liste de tous les diviseurs d'un entier."},"5A12-1":{"url":"/exercices/5e/5A12-1.js","titre":"Primalité ou pas"},"5S13":{"url":"/exercices/5e/5S13.js","titre":"Calculer des fréquences"},"5M20":{"url":"/exercices/5e/5M20.js","titre":"Calculs de volumes"},"5G11":{"url":"/exercices/5e/5G11.js","titre":"Trouver l'image d'un point par une symétrie axiale ou centrale"},"5N11-1":{"url":"/exercices/5e/5N11-1.js","titre":"Tableaux et pourcentages - prix constant"},"5G12-1":{"url":"/exercices/5e/5G12-1.js","titre":"Trouver l'image d'une figure par une symétrie centrale dans un pavage (sortie Latex Impossible)"},"5P10":{"url":"/exercices/5e/5P10.js","titre":"Tableaux et proportionnalité."},"5G20":{"url":"/exercices/5e/5G20.js","titre":"Construire un triangle avec cible auto-corrective"},"5G20-1":{"url":"/exercices/5e/5G20-1.js","titre":"Vocabulaire des triangles"},"5G21-1":{"url":"/exercices/5e/5G21-1.js","titre":"Constructibilité des triangles via les longueurs"},"5L10":{"url":"/exercices/5e/5L10.js","titre":"Écrire une expression littérale"},"5L10-2":{"url":"/exercices/5e/5L10-2.js","titre":"Traduire un programme de calcul par une expression littérale"},"5R11":{"url":"/exercices/5e/5R11.js","titre":"Lire l'abscisse relative d'un point"},"5R20-4":{"url":"/exercices/5e/5R20-4.js","titre":"Résoudre un problème en utilisant une somme algébrique de relatifs."},"5R20-5":{"url":"/exercices/5e/5R20-5.js","titre":"Additions de deux entiers relatifs dans un tableau à double entrée"},"5R20":{"url":"/exercices/5e/5R20.js","titre":"Addition de deux entiers relatifs"},"5L10-1":{"url":"/exercices/5e/5L10-1.js","titre":"Traduire une phrase par une expression"},"5L13-2":{"url":"/exercices/5e/5L13-2.js","titre":"Substitution"},"5CM01":{"url":"/exercices/5e/5CM01.js","titre":"Tables de multiplications"},"5S20":{"url":"/exercices/5e/5S20.js","titre":"Placer un événement sur une échelle de probabilités"},"5A13":{"url":"/exercices/5e/5A13.js","titre":"Décomposition en facteurs premiers"},"5G12":{"url":"/exercices/5e/5G12.js","titre":"Symétrie centrale dans un pavage"},"5G31-1":{"url":"/exercices/5e/5G31-1.js","titre":"Constructibilité des triangles via les angles"},"5N110":{"url":"/exercices/5e/5N110.js","titre":"Variation en pourcentages"},"5L14":{"url":"/exercices/5e/5L14.js","titre":"Calculer la valeur d'une expression littérale"},"5L13":{"url":"/exercices/5e/5L13.js","titre":"Réduire une expression de la forme $ax+bx$"},"5S21":{"url":"/exercices/5e/5S21.js","titre":"Calculer des probabilités dans une expérience aléatoire à une épreuve"},"5R22-2":{"url":"/exercices/5e/5R22-2.js","titre":"Simplifier l'écriture d'une somme de 2 relatifs et calculer"},"5N11-2":{"url":"/exercices/5e/5N11-2.js","titre":"Tableaux et pourcentages - pourcentage constant"},"5R21":{"url":"/exercices/5e/5R21.js","titre":"Soustraction de deux entiers relatifs"},"5G22":{"url":"/exercices/5e/5G22.js","titre":"Déterminer la nature d'une droite remarquable"},"5C11-2":{"url":"/exercices/5e/5C11-2.js","titre":"Traduire une succession des opérations par une expression"},"5R20-2":{"url":"/exercices/5e/5R20-2.js","titre":"Addition à trou de deux entiers relatifs"},"5N20-0":{"url":"/exercices/5e/5N20-0.js","titre":"Résoudre un problème en utilisant des fractions"},"5G31":{"url":"/exercices/5e/5G31.js","titre":"Somme des angles dans un triangle"},"5G30-1":{"url":"/exercices/5e/5G30-1.js","titre":"Déterminer des angles en utilisant les cas d'égalité"},"5C11-1":{"url":"/exercices/5e/5C11-1.js","titre":"Traduire une expression par une phrase"},"3F12-2":{"url":"/exercices/3e/3F12-2.js","titre":"Déterminer l'image d'un nombre par une fonction d'après sa forme algébrique"},"3A13":{"url":"/exercices/3e/3A13.js","titre":"Engrenages"},"3G10-1":{"url":"/exercices/3e/3G10-1.js","titre":"Trouver les coordonnées de l'image d'un point par une transformation du plan"},"3F21-1":{"url":"/exercices/3e/3F21-1.js","titre":"Déterminer une fonction affine"},"3A11-1":{"url":"/exercices/3e/3A11-1.js","titre":"Primalité ou pas - Variante avec les critères de divisibilité par 7 et par 11"},"3P10-1":{"url":"/exercices/3e/3P10-1.js","titre":"Coefficient multiplicateur d'une variation en pourcentage"},"3G12-1":{"url":"/exercices/3e/3G12-1.js","titre":"Trouver l'image d'une figure par une rotation de 90 degrés dans un pavage (sortie Latex Impossible)"},"3A10":{"url":"/exercices/3e/3A10.js","titre":"Division Euclidienne - Diviseurs - Multiples"},"3A12":{"url":"/exercices/3e/3A12.js","titre":"Fractions irréductibles"},"3L14":{"url":"/exercices/3e/3L14.js","titre":"Résoudre une équation produit nul"},"3L13-2":{"url":"/exercices/3e/3L13-2.js","titre":"Equations résolvantes pour le théorème de Thalès"},"3G22":{"url":"/exercices/3e/3G22.js","titre":"Connaître les effets des agrandissements/réductions sur les aires et les volumes"},"3F12":{"url":"/exercices/3e/3F12.js","titre":"Fonctions : Calculs d'images"},"3G23":{"url":"/exercices/3e/3G23.js","titre":"Reconnaître des triangles semblables dans différentes configurations"},"3G30":{"url":"/exercices/3e/3G30.js","titre":"Déterminer une longueur grâce à la trigonométrie"},"3G10-2":{"url":"/exercices/3e/3G10-2.js","titre":"Trouver l'image d'un point par une transformation choisie aléatoirement"},"3G21":{"url":"/exercices/3e/3G21.js","titre":"Contrôler si deux droites sont parallèles"},"3Algo1":{"url":"/exercices/3e/3Algo1.js","titre":"Instruction conditionelle"},"3G10-3":{"url":"/exercices/3e/3G10-3.js","titre":"Construire l'image d'un point par une rotation avec cible auto-corrective"},"3S20":{"url":"/exercices/3e/3S20.js","titre":"Calculer des probabilités dans une expérience aléatoire à deux épreuves"},"3G20":{"url":"/exercices/3e/3G20.js","titre":"Calculer des longueurs avec la propriété de Thalès"},"3S15":{"url":"/exercices/3e/3S15.js","titre":"Calculer des étendues"},"3A11-2":{"url":"/exercices/3e/3A11-2.js","titre":"Décomposition en facteurs premiers d'un entier"},"3L11-1":{"url":"/exercices/3e/3L11-1.js","titre":"Utiliser la double distributivité"},"3F13-1":{"url":"/exercices/3e/3F13-1.js","titre":"Lecture graphique d'images et d'antécédents."},"3L12":{"url":"/exercices/3e/3L12.js","titre":"Factoriser a²-b²"},"3L10-1":{"url":"/exercices/3e/3L10-1.js","titre":"Additionner ou soustraire une expression entre parenthèses"},"3F10":{"url":"/exercices/3e/3F10.js","titre":"Lectures d'images et d'antécédents depuis un tableau de valeurs"},"3L13-1":{"url":"/exercices/3e/3L13-1.js","titre":"Équation du premier degré (utilisant la distributivité)"},"3L11":{"url":"/exercices/3e/3L11.js","titre":"Utiliser la simple distributivité"},"3G11":{"url":"/exercices/3e/3G11.js","titre":"Construire l'image d'un point par une homothetie avec cible auto-corrective"},"3G43":{"url":"/exercices/3e/3G43.js","titre":"Calculs de volumes"},"3F12-3":{"url":"/exercices/3e/3F12-3.js","titre":"Compléter un tableau de valeurs"},"3A11":{"url":"/exercices/3e/3A11.js","titre":"Primalité ou pas"},"3P10":{"url":"/exercices/3e/3P10.js","titre":"Variations en pourcentage"},"3G12":{"url":"/exercices/3e/3G12.js","titre":"Trouver l'image d'une figure par une rotation dans un pavage"},"3L11-4":{"url":"/exercices/3e/3L11-4.js","titre":"Factoriser une expression"},"3L15":{"url":"/exercices/3e/3L15.js","titre":"Résoudre une équation du second degré"},"3F1-act":{"url":"/exercices/3e/3F1-act.js","titre":"Fonctions : Notion et vocabulaire"},"3L10":{"url":"/exercices/3e/3L10.js","titre":"Donner l'opposé d'une expression"},"3F12-4":{"url":"/exercices/3e/3F12-4.js","titre":"Lire l'image d'un nombre à partir d'un graphique"},"3F13-2":{"url":"/exercices/3e/3F13-2.js","titre":"Spécial escape game"},"3L12-1":{"url":"/exercices/3e/3L12-1.js","titre":"Développer (a-b)(a+b)"},"3L11-3":{"url":"/exercices/3e/3L11-3.js","titre":"Utiliser la distributivité (simple ou double) et réduire"},"3A11-3":{"url":"/exercices/3e/3A11-3.js","titre":"Compter/lister les diviseurs d'un entier à partir de sa décomposition en facteurs premiers."},"3G20-1":{"url":"/exercices/3e/3G20-1.js","titre":"Problèmes avec le théorème de Thalès"},"3G31":{"url":"/exercices/3e/3G31.js","titre":"Déterminer un angle grâce à la trigonométrie"},"3L11-5":{"url":"/exercices/3e/3L11-5.js","titre":"Calcul mental et calcul littéral"},"3L14-1":{"url":"/exercices/3e/3L14-1.js","titre":"Résoudre une équation produit nul (niveau 2)"},"3G20-2":{"url":"/exercices/3e/3G20-2.js","titre":"Déterminer une longueur avec la propriété de Thales (MG32)"},"3L13":{"url":"/exercices/3e/3L13.js","titre":"Équation du premier degré"},"3F13":{"url":"/exercices/3e/3F13.js","titre":"Lire les antécédents d'un nombre à partir d'un graphique"},"3L11-2":{"url":"/exercices/3e/3L11-2.js","titre":"Réduire une expression"},"3F21":{"url":"/exercices/3e/3F21.js","titre":"Déterminer une fonction linéaire"},"2L11":{"url":"/exercices/2e/2L11.js","titre":"Factoriser avec les identités remarquables"},"2N20":{"url":"/exercices/2e/2N20.js","titre":"Déterminer le plus petit ensemble de nombres dans lequel le nombre proposé appartient"},"2G10":{"url":"/exercices/2e/2G10.js","titre":"Utiliser la distance entre deux points dans un repère orthonormé"},"2N10":{"url":"/exercices/2e/2N10.js","titre":"Existence d'une racine carrée"},"2N12":{"url":"/exercices/2e/2N12.js","titre":"Appliquer la double distributivité avec les racines carrées"},"2N21":{"url":"/exercices/2e/2N21.js","titre":"Déterminer la parité d'une expression"},"2N24":{"url":"/exercices/2e/2N24.js","titre":"Associer un intervalle de  $\\mathbb{R}$ à une inéquation et son schéma sur une droite graduée"},"2N25":{"url":"/exercices/2e/2N25.js","titre":"Utiliser et comprendre les symboles $\\cup $ et $\\cap $ avec les intervalles de $\\mathbb{R}$"},"2N12-1":{"url":"/exercices/2e/2N12-1.js","titre":"Développer les identités remarquables avec des racines carrées"},"2N11-1":{"url":"/exercices/2e/2N11-1.js","titre":"Simplifier une somme de racines carrées"},"2N22":{"url":"/exercices/2e/2N22.js","titre":"Utiliser la notion de valeur absolue d'une quantité"},"2N11":{"url":"/exercices/2e/2N11.js","titre":"Ecrire le nombre proposé sous la forme $a\\sqrt{b}$"},"2L10":{"url":"/exercices/2e/2L10.js","titre":"Développer avec les identités remarquables"},"2N10-1":{"url":"/exercices/2e/2N10-1.js","titre":"Connaître les propriétés calculatoires des racines carrées"},"2G11":{"url":"/exercices/2e/2G11.js","titre":"Déterminer les coordonnées milieu d’un segment dans un repère"},"2G12":{"url":"/exercices/2e/2G12.js","titre":"Déterminer la nature d'un polygone."},"2N23":{"url":"/exercices/2e/2N23.js","titre":"Résoudre une équation avec des valeurs absolues"},};
 
 export function menuDesExercicesDisponibles(){
 let liste_des_exercices_disponibles = tridictionnaire(dictionnaireDesExercices);
